LibGfx/JBIG2: Move GenericContexts to JBIG2Shared.h
No behavior change.
