+++ Userland/Libraries/LibGfx/ImageFormats/JBIG2Loader.cpp
@@ -1006,28 +1006,8 @@ struct GenericRegionDecodingInputParameters {
     QMArithmeticDecoder* arithmetic_decoder { nullptr };
 };
 
-struct GenericContexts {
-    GenericContexts(u8 template_)
-    {
-        contexts.resize(1 << number_of_context_bits_for_template(template_));
-    }
-
-    Vector<QMArithmeticCoderContext> contexts; // "GB" (+ binary suffix) in spec.
-
-private:
-    static u8 number_of_context_bits_for_template(u8 template_)
-    {
-        if (template_ == 0)
-            return 16;
-        if (template_ == 1)
-            return 13;
-        VERIFY(template_ == 2 || template_ == 3);
-        return 10;
-    }
-};
-
 // 6.2 Generic region decoding procedure
-static ErrorOr<NonnullOwnPtr<BilevelImage>> generic_region_decoding_procedure(GenericRegionDecodingInputParameters const& inputs, Optional<GenericContexts>& maybe_contexts)
+static ErrorOr<NonnullOwnPtr<BilevelImage>> generic_region_decoding_procedure(GenericRegionDecodingInputParameters const& inputs, Optional<JBIG2::GenericContexts>& maybe_contexts)
 {
     if (inputs.is_modified_modified_read) {
         dbgln_if(JBIG2_DEBUG, "JBIG2ImageDecoderPlugin: MMR image data");
@@ -1722,14 +1702,14 @@ static ErrorOr<Vector<NonnullRefPtr<Symbol>>> symbol_dictionary_decoding_procedu
     Optional<FixedMemoryStream> stream;
     Optional<BigEndianInputBitStream> bit_stream;
     Optional<QMArithmeticDecoder> decoder;
-    Optional<GenericContexts> generic_contexts;
+    Optional<JBIG2::GenericContexts> generic_contexts;
     Optional<SymbolContexts> symbol_contexts;
     if (inputs.uses_huffman_encoding) {
         stream = FixedMemoryStream { data };
         bit_stream = BigEndianInputBitStream { MaybeOwned { stream.value() } };
     } else {
         decoder = TRY(QMArithmeticDecoder::initialize(data));
-        generic_contexts = GenericContexts { inputs.symbol_template };
+        generic_contexts = JBIG2::GenericContexts { inputs.symbol_template };
         symbol_contexts = SymbolContexts {};
     }
 
@@ -2087,7 +2067,7 @@ struct GrayscaleInputParameters {
     QMArithmeticDecoder* arithmetic_decoder { nullptr };
 };
 
-static ErrorOr<Vector<u64>> grayscale_image_decoding_procedure(GrayscaleInputParameters const& inputs, ReadonlyBytes data, Optional<GenericContexts>& contexts)
+static ErrorOr<Vector<u64>> grayscale_image_decoding_procedure(GrayscaleInputParameters const& inputs, ReadonlyBytes data, Optional<JBIG2::GenericContexts>& contexts)
 {
     VERIFY(inputs.bpp < 64);
 
@@ -2186,7 +2166,7 @@ struct HalftoneRegionDecodingInputParameters {
 };
 
 // 6.6 Halftone Region Decoding Procedure
-static ErrorOr<NonnullOwnPtr<BilevelImage>> halftone_region_decoding_procedure(HalftoneRegionDecodingInputParameters const& inputs, ReadonlyBytes data, Optional<GenericContexts>& contexts)
+static ErrorOr<NonnullOwnPtr<BilevelImage>> halftone_region_decoding_procedure(HalftoneRegionDecodingInputParameters const& inputs, ReadonlyBytes data, Optional<JBIG2::GenericContexts>& contexts)
 {
     // 6.6.5 Decoding the halftone region
     // "1) Fill a bitmap HTREG, of the size given by HBW and HBH, with the HDEFPIXEL value."
@@ -2298,7 +2278,7 @@ struct PatternDictionaryDecodingInputParameters {
 };
 
 // 6.7 Pattern Dictionary Decoding Procedure
-static ErrorOr<Vector<NonnullRefPtr<Symbol>>> pattern_dictionary_decoding_procedure(PatternDictionaryDecodingInputParameters const& inputs, ReadonlyBytes data, Optional<GenericContexts>& contexts)
+static ErrorOr<Vector<NonnullRefPtr<Symbol>>> pattern_dictionary_decoding_procedure(PatternDictionaryDecodingInputParameters const& inputs, ReadonlyBytes data, Optional<JBIG2::GenericContexts>& contexts)
 {
     // Table 27 – Parameters used to decode a pattern dictionary's collective bitmap
     GenericRegionDecodingInputParameters generic_inputs;
@@ -2991,9 +2971,9 @@ static ErrorOr<void> decode_pattern_dictionary(JBIG2LoadingContext&, SegmentData
     // Done!
 
     // "2) As described in E.3.7, reset all the arithmetic coding statistics to zero."
-    Optional<GenericContexts> contexts;
+    Optional<JBIG2::GenericContexts> contexts;
     if (!uses_mmr)
-        contexts = GenericContexts { hd_template };
+        contexts = JBIG2::GenericContexts { hd_template };
 
     // "3) Invoke the pattern dictionary decoding procedure described in 6.7, with the parameters to the pattern
     //     dictionary decoding procedure set as shown in Table 35."
@@ -3082,9 +3062,9 @@ static ErrorOr<void> decode_immediate_halftone_region(JBIG2LoadingContext& conte
         return Error::from_string_literal("JBIG2ImageDecoderPlugin: Halftone segment without patterns");
 
     // "3) As described in E.3.7, reset all the arithmetic coding statistics to zero."
-    Optional<GenericContexts> contexts;
+    Optional<JBIG2::GenericContexts> contexts;
     if (!uses_mmr)
-        contexts = GenericContexts { template_used };
+        contexts = JBIG2::GenericContexts { template_used };
 
     // "4) Invoke the halftone region decoding procedure described in 6.6, with the parameters to the halftone
     //     region decoding procedure set as shown in Table 36."
@@ -3196,9 +3176,9 @@ static ErrorOr<void> decode_immediate_generic_region(JBIG2LoadingContext& contex
     // "1) Interpret its header, as described in 7.4.6.1"
     // Done above.
     // "2) As described in E.3.7, reset all the arithmetic coding statistics to zero."
-    Optional<GenericContexts> contexts;
+    Optional<JBIG2::GenericContexts> contexts;
     if (!uses_mmr)
-        contexts = GenericContexts { arithmetic_coding_template };
+        contexts = JBIG2::GenericContexts { arithmetic_coding_template };
 
     // "3) Invoke the generic region decoding procedure described in 6.2, with the parameters to the generic region decoding procedure set as shown in Table 37."
     GenericRegionDecodingInputParameters inputs;
